- Some warning fixes. - Added Planeshift.Client.Loading.UnloadLast option to psclient.cfg which   alters whether unloading is done before or after map loading. Set to true   by default, maps are loaded before old are unloaded (faster for machines   with enough ram to not swap, possibly slower for those who swap). Will   add GUI option in future.

+++ trunk/src/common/util/pspath.cpp
@@ -197,7 +197,7 @@ float psPath::Distance(psWorld * world, iEngine *engine,csVector3& pos, iSector 
         if (dist < 0 || d < dist)
         {
             dist = d;
-            idx = i;
+            idx = (int)i;
         }
     }
     if (dist >= 0.0)
